Check if file is avif before converting

+++ bin/lib.js
@@ -17,15 +17,30 @@ import sharp from 'sharp';
 const convertToJPEG = (file = '', { quality = 100, isWillOverwrite = false }) => {
   if (!file) return;
   const jpgFile = `${file.substring(0, file.lastIndexOf("."))}.jpg`;
-  sharp(`${file}`)
-    .jpeg({ quality: quality })
-    .toFile(`${jpgFile}`)
-    .then(() => {
-      console.log(`=> ${jpgFile} (quality: ${quality} %)`);
-      if (isWillOverwrite) {
-        fs.unlinkSync(`${file}`);
-      }
-    })
+  
+  // Read the first 32 bytes of the file to check if it has a valid AVIF header
+  const fd = fs.openSync(file, 'r');
+  const buffer = Buffer.alloc(32);
+  fs.readSync(fd, buffer, 0, 32, 0);
+  const hasAvifHeader = buffer.slice(4, 8).toString() === 'ftyp' && buffer.slice(8, 12).toString() === 'avif';
+  fs.closeSync(fd);
+
+  if (hasAvifHeader) {
+    sharp(`${file}`)
+      .jpeg({ quality: quality })
+      .toFile(`${jpgFile}`)
+      .then(() => {
+        console.log(`=> ${jpgFile} (quality: ${quality} %)`);
+        if (isWillOverwrite) {
+          fs.unlinkSync(`${file}`);
+        }
+      })
+      .catch((err) => {
+        console.error(`Error converting ${file} to JPEG: ${err}`);
+      });
+  } else {
+    console.log(`Skipping ${file}: file format not supported.`);
+  }
 }
 
 /**
